requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
ca-certificates=2020.12.8=h06a4308_0
certifi=2020.12.5=py38h06a4308_0
data-science-types=0.2.22=pypi_0
iniconfig=1.1.1=py_0
intel-openmp=2020.2=254
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
more-itertools=8.6.0=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
openssl=1.1.1i=h27cfd23_0
packaging=20.8=pyhd3eb1b0_0
pandas=1.1.5=py38ha9443f7_0
pip=20.3.3=py38h06a4308_0
pluggy=0.13.1=py38_0
py=1.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=py_0
pytest=6.2.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
pytz=2020.4=pyhd3eb1b0_0
pyyaml=5.3.1=pypi_0
readline=8.0=h7b6447c_0
setuptools=51.0.0=py38h06a4308_2
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
